+++ packages/orq-rc/src/orq_ai_sdk/models/get_v2_human_eval_setsop.py
@@ -100,7 +100,7 @@ class GetV2HumanEvalSetsResponseBody2(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-12T10:25:48.200Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-13T03:39:40.773Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -185,7 +185,7 @@ class GetV2HumanEvalSetsResponseBody1(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-12T10:25:48.200Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-13T03:39:40.773Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")