+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KP0KM9AR84GMZ9KN38RAMTZT"
+        "01KP2ESBP3YVAK66KSWTYKQ2EJ"
     )
     r"""The unique identifier of the data source"""