+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KP0KM9AV4GX79M4VKWPTEWJ4"
+        "01KP2ESBP578A3P7FGFNMD5A34"
     )
     r"""The unique identifier of the data source"""