+++ packages/orq-rc/RELEASES.md
@@ -7069,3 +7069,13 @@ Based on:
 - [python v4.8.0-rc.6] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.6] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.6 - packages/orq-rc
+
+## 2026-04-13 03:42:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.3 (2.879.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.7] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.7 - packages/orq-rc